+++ extension.js
@@ -8,56 +8,85 @@ function activate(context) {
 
     // Use the console to output diagnostic information (console.log) and errors (console.error)
     // This line of code will only be executed once when your extension is activated
-    console.log('Congratulations, your extension "r-pack-build" is now active!');
+    console.log('Congratulations, your extension "r-vscode-tools" is now active!');
 
-    // The command has been defined in the package.json file
-    // Now provide the implementation of the command with  registerCommand
-    // The commandId parameter must match the command field in package.json
-    var disposable = vscode.commands.registerCommand('extension.checkRpackage', function () {
+    // Command to check R package
+    context.subscriptions.push(vscode.commands.registerCommand('extension.checkRpackage', function () {
         // The code you place here will be executed every time your command is executed
         var shell = require('shelljs');
         console.log(vscode.workspace.rootPath);
         shell.cd(vscode.workspace.rootPath);
         var buildCommand = 'R CMD build ' + vscode.workspace.rootPath;
         var checkCommand = 'R CMD check *.tar.gz';
-        var step2 = true;
         var outputChannel = vscode.window.createOutputChannel("R");
         outputChannel.show();
-        shell.exec(buildCommand, 
-            function(code, stdout, stderr) {
-                console.log(buildCommand);
-                console.log(stdout);
-                console.log(code);
-                outputChannel.append(buildCommand);
-                outputChannel.append(stdout);
-                if(code) {
-                    step2 = false;
-                    vscode.window.showInformationMessage('Building tarball failed!');
-                    console.log('Build failed!');
+        outputChannel.append(buildCommand + "\n");
+        let p1 = new Promise((resolve, reject) => {
+            shell.exec(buildCommand, function (code, stdout, stderr) {
+                if (code) {
                     console.log(stderr);
                     outputChannel.append(stderr);
-                }
-            });
-        if(step2) {
-            shell.exec(checkCommand, function (cCode, cStdout, cStderr) {
-                console.log(checkCommand);
-                console.log(cStdout);
-                outputChannel.append(checkCommand);
-                outputChannel.append(cStdout);
-                if (cCode) {
-                    vscode.window.showInformationMessage('Check failed!');
-                    console.log('Check failed!');
-                    console.log(cStderr);
-                    outputChannel.append(cStderr);
+                    vscode.window.setStatusBarMessage("Build failed!");
+                    reject();
                 } else {
-                    vscode.window.showInformationMessage('Check succeeded!');
+                    outputChannel.append("Done!\n");
+                    vscode.window.setStatusBarMessage('Build succeeded!', 2000);
+                    resolve();
                 }
             });
-        }
-        // Display a message box to the user
-    });
+        });
+        p1.then(function () {
+            outputChannel.append(checkCommand + "\n");
+            shell.exec(checkCommand, function (code, stdout, stderr) {
+                if (code) {
+                    outputChannel.append(stderr);
+                    outputChannel.append("Failed!");
+                } else {
+                    outputChannel.append("Done!");
+                    vscode.window.setStatusBarMessage('Check succeeded!', 5000);
+                }
+            });
+        });
+    }));
 
-    context.subscriptions.push(disposable);
+    // Command to install an R package directory
+    context.subscriptions.push(vscode.commands.registerCommand('extension.installRpackage', function () {
+        // The code you place here will be executed every time your command is executed
+        var shell = require('shelljs');
+        console.log(vscode.workspace.rootPath);
+        shell.cd(vscode.workspace.rootPath);
+        var buildCommand = 'R CMD build ' + vscode.workspace.rootPath;
+        var installCommand = 'R CMD INSTALL *.tar.gz';
+        var outputChannel = vscode.window.createOutputChannel("R");
+        outputChannel.show();
+        outputChannel.append(buildCommand + "\n");
+        let p1 = new Promise((resolve, reject) => {
+            shell.exec(buildCommand, function(code, stdout, stderr) {
+                if(code) {
+                    console.log(stderr);
+                    outputChannel.append(stderr);
+                    vscode.window.setStatusBarMessage("Build failed!");
+                    reject();
+                } else {
+                    outputChannel.append("Done!\n");
+                    vscode.window.setStatusBarMessage('Build succeeded!', 2000);
+                    resolve();
+                }
+            });
+        });
+        p1.then(function() {
+            outputChannel.append(installCommand + "\n");
+            shell.exec(installCommand, function(code, stdout, stderr){
+                if(code) {
+                    outputChannel.append(stderr);
+                    outputChannel.append("Failed!");
+                } else {
+                    outputChannel.append("Done!");
+                    vscode.window.setStatusBarMessage('Install succeeded!', 5000);
+                }
+            });
+        });
+    }));
 }
 exports.activate = activate;