+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # Change Log
-All notable changes to the "r-check" extension will be documented in this file.
+All notable changes to the "r-vscode-tools" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
-- Initial release
+## 0.0.1
+- Add command `R: Check package`
+
+## 0.0.2
+- Add command `R: Install package directory`
+- Restructure JS code using Promises
+- Display information in the status bar
+- Less information in the output channel